정아

+++ TravelingProject/src/main/java/com/human/springboot/LDTO.java
@@ -29,8 +29,8 @@ public class LDTO {
 	String place_date;
 	String place_open;
 	int place_like;
-	int place_lat;
-	int place_lng;
+	double place_lat;
+	double place_lng;
 	
 	
 	//city
